+++ ocs_sdk/tests/apis/test_api.py
@@ -27,18 +27,3 @@ class TestAPI(FakeAPITestCase, unittest.TestCase):
 
     def test_get_api_url(self):
         self.assertEqual(SimpleAPI().get_api_url(), 'http://localhost')
-
-    def test_safe_query(self):
-        api = API()
-        api.base_url = 'http://localhost/'
-
-        self.fake_endpoint(api, 'users/', status=500)
-
-        self.assertIsNone(
-            api.safe_query(api.query().users.get, http_status_caught=[500]),
-        )
-
-        self.assertRaises(
-            slumber.exceptions.SlumberHttpBaseException,
-            api.safe_query, api.query().users.get
-        )